requirements.txt
Django==2.1
django-widget-tweaks==1.4.2
python-keystoneclient==3.17.0
python-dateutil==2.7.3
django-charsleft-widget==0.1.6
django-phonenumber-field==2.0.0
python-decouple==3.1
